feat(target) support for powershell's invoke-restmethod
