+++ src/targets/powershell/restmethod.js
@@ -1,0 +1,10 @@
+'use strict'
+
+module.exports = require('./common')('Invoke-RestMethod');
+
+module.exports.info = {
+  key: 'restmethod',
+  title: 'Invoke-RestMethod',
+  link: 'https://docs.microsoft.com/en-us/powershell/module/Microsoft.PowerShell.Utility/Invoke-RestMethod',
+  description: 'Powershell Invoke-RestMethod client'
+}
